+++ CHANGELOG.md
@@ -7,6 +7,7 @@ NeoForge-1.21.1-21.1.230
 - Scarstone may now be crushed for strange crystals [(\#186)](https://github.com/EnigmaticaModpacks/EnigmaticSkies/issues/186)
 - Improve performance of the Animal Senescence system [(\#189)](https://github.com/EnigmaticaModpacks/EnigmaticSkies/issues/189)
 - Various decorative mods now accept stick, plank and/or glass tags where appropriate [(\#192)](https://github.com/EnigmaticaModpacks/EnigmaticSkies/issues/192)
+- Fermentation Starter moved from the Industrial Cauldron to the Composter [(\#192)](https://github.com/EnigmaticaModpacks/EnigmaticSkies/issues/192)
 
 #### 🦟 Bugs Fixed